requirements.txt
pip==21.3.1
setuptools==58.5.3
discord==1.7.3
googlemaps~=4.5.3
python-dotenv~=0.19.1